# backend/docs/swagger.yaml
definitions:
  http_common.Error:
    properties:
      code:
        type: string
      field:
        type: string
      message:
        type: string
    type: object
  http_common.HttpResponse-any:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/http_common.Error'
        type: array
      success:
        type: boolean
    type: object
  http_common.HttpResponse-string:
    properties:
      data:
        type: string
      errors:
        items:
          $ref: '#/definitions/http_common.Error'
        type: array
      success:
        type: boolean
    type: object
  model.UserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
info:
  contact: {}
paths:
  /users/login:
    post:
      consumes:
      - application/json
      description: Login to account
      parameters:
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_common.HttpResponse-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_common.HttpResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_common.HttpResponse-any'
      summary: Login
      tags:
      - Users
  /users/register:
    post:
      consumes:
      - application/json
      description: Register to account
      parameters:
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/http_common.HttpResponse-string'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http_common.HttpResponse-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http_common.HttpResponse-any'
      summary: Register
      tags:
      - Users
swagger: "2.0"